# Sem4/OOP/Lab7/docs/swagger.yaml
definitions:
  models.AddProblemsInfo:
    properties:
      problems:
        items:
          type: integer
        type: array
      session_id:
        type: integer
    type: object
  models.AvailabilitySlot:
    properties:
      candidate_id:
        type: integer
      end_time:
        type: string
      id:
        type: integer
      interviewer_id:
        type: integer
      is_booked:
        type: boolean
      start_time:
        type: string
    type: object
  models.IdSetter:
    properties:
      id:
        type: integer
    type: object
  models.InterwieweProblem:
    properties:
      grade:
        type: integer
      id:
        type: integer
      text:
        type: string
      theme:
        type: string
    type: object
  models.LoginInfo:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  models.RegisterInfo:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  models.ResultIndo:
    properties:
      hard_skils:
        type: string
      session_id:
        type: integer
      soft_skils:
        type: string
      to_hire:
        type: string
    type: object
  models.SessionRegisterInfo:
    properties:
      candidate_id:
        type: integer
      interwiwer_id:
        type: integer
    type: object
info:
  contact: {}
  title: Interwier
  version: "1.0"
paths:
  /api/candidates/{candidateId}/slots/{slotId}/book:
    post:
      description: Кандидат бронирует слот по его ID
      parameters:
      - description: Candidate ID
        in: path
        name: CandidateId
        required: true
        type: integer
      - description: Slot ID
        in: path
        name: SlotId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - PASETOAuth: []
      summary: Забронировать слот
      tags:
      - slots
  /api/candidates/{candidateId}/slots/{slotId}/unbook:
    post:
      description: Кандидат отменяет бронирование слота по его ID
      parameters:
      - description: Candidate ID
        in: path
        name: CandidateId
        required: true
        type: integer
      - description: Slot ID
        in: path
        name: SlotId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - PASETOAuth: []
      summary: Отменить бронирование слота
      tags:
      - slots
  /api/candidates/{id}:
    delete:
      description: Удаляет кандидата по ID в защищённой зоне
      parameters:
      - description: Candidate ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - PASETOAuth: []
      summary: Удалить кандидата (требует PASETO токен)
      tags:
      - candidates
  /api/candidates/{id}/slots:
    get:
      description: Возвращает список доступных слотов для записи по ID интервьюера
      parameters:
      - description: Interviewer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.AvailabilitySlot'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить доступные слоты интервьюера
      tags:
      - slots
  /api/problems/{id}:
    delete:
      description: Удаляет задачу по ID
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Problem deleted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - PASETOAuth: []
      summary: Удалить задачу (требует PASETO токен)
      tags:
      - problems
    get:
      description: Возвращает задачу по её ID
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InterwieweProblem'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - PASETOAuth: []
      summary: Получить задачу по ID (требует PASETO токен)
      tags:
      - problems
  /api/problems/add:
    post:
      consumes:
      - application/json
      description: Создаёт новую задачу для интервью
      parameters:
      - description: Problem data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.InterwieweProblem'
      produces:
      - application/json
      responses:
        "201":
          description: Problem added
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - PASETOAuth: []
      summary: Добавить задачу (требует PASETO токен)
      tags:
      - problems
  /api/problems/all:
    get:
      description: Возвращает список всех задач
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.InterwieweProblem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - PASETOAuth: []
      summary: Получить все задачи (требует PASETO токен)
      tags:
      - problems
  /api/session/{id}:
    delete:
      parameters:
      - description: Candidate ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - PASETOAuth: []
      summary: Удалить сессию (требует PASETO токен)
      tags:
      - session
  /api/session/{id}/start:
    post:
      parameters:
      - description: session ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - PASETOAuth: []
      summary: Начать сессию
      tags:
      - session
  /api/session/{id}/stop:
    post:
      parameters:
      - description: session ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - PASETOAuth: []
      summary: Остановить сессию
      tags:
      - session
  /api/session/all:
    get:
      consumes:
      - application/json
      description: Получить все сессии в защищённой зоне
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - PASETOAuth: []
      summary: Получить все сессии (требует PASETO токен)
      tags:
      - session
  /api/session/create:
    post:
      consumes:
      - application/json
      description: Создаёт сессию в защищённой зоне
      parameters:
      - description: session info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SessionRegisterInfo'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - PASETOAuth: []
      summary: создать сессию (требует PASETO токен)
      tags:
      - session
  /api/session/get:
    post:
      consumes:
      - application/json
      description: Получить сесси.ю в защищённой зоне
      parameters:
      - description: session id
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.IdSetter'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - PASETOAuth: []
      summary: Получить сессию(требует PASETO токен)
      tags:
      - session
  /api/session/problems:
    post:
      parameters:
      - description: session problems info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AddProblemsInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - PASETOAuth: []
      summary: Добавить задачи в сессию
      tags:
      - session
  /api/session/result:
    post:
      parameters:
      - description: session result info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ResultIndo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - PASETOAuth: []
      summary: Добавть результат сесии
      tags:
      - session
  /api/slots/{id}:
    delete:
      parameters:
      - description: Slot ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - PASETOAuth: []
      summary: Удалить бронь времени (требует PASETO токен)
      tags:
      - slot
  /api/slots/{id}/book:
    post:
      parameters:
      - description: session ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - PASETOAuth: []
      summary: Забронировать время
      tags:
      - slot
  /api/slots/{id}/unbook:
    post:
      parameters:
      - description: session ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - PASETOAuth: []
      summary: Освободить время
      tags:
      - slot
  /api/slots/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: session info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AvailabilitySlot'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - PASETOAuth: []
      summary: Создать времной слот (требует PASETO токен)
      tags:
      - slot
  /login:
    post:
      consumes:
      - application/json
      description: Login user and return pasetto tocken
      parameters:
      - description: Sign in info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Login user
      tags:
      - log in
  /register:
    post:
      consumes:
      - application/json
      description: Returns nil error of success
      parameters:
      - description: Sign in info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Register user
      tags:
      - sign in
securityDefinitions:
  PASETOAuth:
    description: Введите "Bearer <PASETO токен>". Получите токен на /login и прикрепите
      его.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"